requirements.txt
aiogram==3.21.0
asyncpg==0.30.0
aiofiles==24.1.0
python-dotenv==1.0.1
Pillow==10.4.0
matplotlib==3.9.4
pandas==2.3.1